+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.13.0 (2024-xx-xx)
 
 - Improvement: Messages On Behalf Of Channels feature
+- Improvement: updated tooltip for using the voteban command
 - Other: internal improvements
 
 ## 1.12.0 (2023-12-31)